Move viewable stuff to PacketViewableUtils
